+++ packages/ai-eval/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @gtcx/ai-eval
 
+## 0.1.3
+
+### Patch Changes
+
+- Updated dependencies
+  - @gtcx/types@3.1.4
+
+## 0.1.2
+
+### Patch Changes
+
+- Updated dependencies [fe01c1a]
+  - @gtcx/types@3.1.3
+
 ## 0.1.1
 
 ### Patch Changes